Fix event and tally_state data
This patch adds a liquibase changeset to fix the data in the events and
tally_state tables caused by a failed deployment. The changeset is
intended to be run in a single transaction.

The changeset will:
1. Set the record_date in the events.data JSON column to the record's record_date
   when the JSON's record_date is NULL.
2. Set the tally_state.latest_event_record_date to '2024-04-01T01:00Z'
   when the latest_event_record_date is null.
3. Add a non-null constraint on the tally_state.latest_event_record_date
   column

NOTE: Due to the offset of job runs, an Event's record_date will be
      approximately 1.5h after the event timestamp. Therefore, we
      set the latest_event_record_date to '2024-04-01T01:00Z' to
      ensure that we do not include any events with timestamps in
      March.