SWATCH-2388 Add unit test
HBI: For high watermark or nightly tally data we store in tally_snapshots and tally_measurements. We don't store per host for traditional it's the max value always.

Cost or others: We store hourly tally in instance_monthly_total for payg or non-payg. We also store in tally_measurements and tally_snapshot but, we store by hosts and month.

Signed-off-by: Kartik Shah <karshah@redhat.com>