SWATCH-2275: Add the product-client module for Quarkus
- Create the product-client for Quarkus.
- Copy the ProductService, ProductDataSource, StubProductApi and ProductApiFactory classes into the new module

Since we don't have the concept of FactoryBean in Quarkus and also the rest clients are injected via client proxies in Quarkus. I figured out to inject the ProductApi instances via the qualifier ProductClient annotation. This way, the ProductApiFactory plays to produce either the stub product api or the generated product api.

Note that to fully test these changes, I had to do some minimal configuration in the swatch-contracts.