SWATCH-2444: Fix bug with multiple azure API credentials  (#3246)
<!-- Replace XXXX with the issue number. Issue will be auto-linked -->
Jira issue: SWATCH-2444

## Description
<!-- Provide a description of this PR. Try to provide answers to "what",
"how",
and "why" -->
When we have multiple sets of api credentials to use against azure, the
code was calling the api with both credentials every time. This code
change makes it so we only try the next set of credentials if the first
one fails instead of continuing to try more credentials.

To test this you need to use credentials from stage, updated to have
another invalid entry.