Allow Kafka in Quarkus apps to default to the system truststore
Passing the empty string as the location of the truststore to Kafka
results in an error since that location does not exist.  If one wants to
use the system truststore (which we do in production and stage), one
must not set the location at all (leaving it null).

See SWATCH-2464 and SWATCH-2950