Capture special pricing flag from product service
A new field got added to SKUs (SPECIAL_PRICING_FLAG) and we want to detect if it's set to MIGRATION_OFFERING, which will indicate that particular sku is for a third-party migrated offering.  Making that differentiation will allow us to tally 3rd party migrated instances separately.  With this commit, swatch will now consume that flag during Offering sync and persist that information in the database.

Co-authored-by: Jose Carvajal <josecarvajalhilario@gmail.com>