SWATCH-2465: Do not filter by metric if not set in the Instances API
Before, we were using the default (first) metric ID by product tag if the user does not specify any.
After these changes, we'll be getting all the metrics if the user does not specify any.
To avoid getting duplicates, if no metric ID filter is used, we're going to filter by all the valid metrics in the product tag.
This logic affects the Instances API and also the Export Instances export, so I covered this change in both existing tests.