SWATCH-2483: Fix export subscriptions service to filter by payg
The unlimited subscriptions query is used to populate the total capacity of the skus, and since we are not gathering this information in the export service, we don't need it.
To retrieve the payg subscriptions, the difference is how to filter by product (it's being done one way if the product is non payg, and in a different way if the product is payg). I have a fix for this to retrieve both.