Remove dependency of EventKey from swatch-metrics
Maintaining a set of Events that had been set and checking against EventKey was legacy code from when events got immediately persisted to the database.

Event ingestion is able to dedup events now, so we don't need to maintain that set or check against it prior to putting an Event on the service-instance-ingress topic.