Event amendment refactor
An event's potential overlaps (conflicts) are determined
based on an EventKey. This key is based on the event's org,
instance-id and timestamp, all of which identify the usage
that an event applies to. All non relevant properties have
been removed from the key to ensure that usage is only based
on a specific instance/timestamp, no matter where it was
reported from.

An incoming Event is first normalized (flattened) into a
set of Events that have a single tag/measurement combo.
Each of which are then resolved individually, and amendments
are made to each to ensure that all usage at the specified
timestamp is first zeroed out before the new usage is applied.

And event is amended, if the measurement values of two compared
events are different, OR if the hardwareType, sla, usage,
billing provider, or billingProviderId, is different.

NOTES:
* Removed usage of EventKey from the EventController.
* Removed eventType and eventSource from EventKey since they do
  not relate to identifying the usage that the event contributes to.
* Consider the product tags of events when determining
  which amendments are required to resolve an incoming
  conflict.
* Consider incoming messages as conflicts after they
  have been processed so they are considered by other
  incoming events.
* Use LinkedHashMap when grouping events by key in order
  to maintain the order that they are processed.