Add product tag verification to validateServiceInstanceEvent
Added more logic to ensure that product tag is matching a product config definition, and to fail validation if there isn't a defined product tag or if the event message has defined product tags that don't match the ones from the lookup.

If there is not a matching product tag, the event is considered invalid and will not get persisted.