SWATCH-2586: Change sourcePartner type to string
Before, the sourcePartner was an enum with a restricted set of values (aws_marketplace and azure_marketplace). 
The problem is that the IT Partner Gateway uses a string and as far as I could see, there are much more valid values like gcp_marketplace, ibm_marketplace, ... And it could be more in the future. 
Therefore, we should use a free string instead and check the marketplace we support as we already do nowadays.