SWATCH-2594: Ensure ordering of streams when collating hosts
The problem is that when querying the hosts in stage, we're retrieving the following hosts at this order:

First iteration:
- Inventory Host with InventoryId=YYY
- Swatch Host with InventoryId=AAA

Second iteration: 
- Inventory Host with InventoryId=ANY
- Swatch Host with InventoryId=YYY (it should be found in the first iteration!)

I don't see how we could possible resolve the issue by tweaking the order by statement in the HostRepository.streamHbiHostsByOrgId since the InventoryId is ordered alphabetically, so the above is a realistic scenario).

The only solution that I can think of is to sorting the streams using the SortKey.