Only touch non-PAYG buckets during nightly tally
Since a host's `isMetered` status is determined based on buckets existing for a isPaygEligible product tag, we don't want to incorrectly clear those buckets during host reconciliation.  If we don't enforce `applicableProducts` throughout the host reconciliation that happens during nightly tally.  It's possible for a host that's metered via prometheus or CostMgmt to also be reported by HBI.  When this happens, if we don't enforce/limit `applicableProducts` to non-PAYG tags during nightly host reconciliation, we would wipe out the tally buckets that were indicating the host is metered.  Which would result in the next nightly tally (given that no more hourly snaphots/buckets get created..i.e. that system gets shut down) to incorrectly consider that host as `isMetered: false`, and make a snapshot associated with incorrect product tags.

Co-authored-by: Kevin Howell <khowell@redhat.com>
Co-authored-by: Lindsey Burnett <lburnett@redhat.com>