Do not terminate subscriptions repeatedly when quantity changes
When the quantity changes, we terminate a subscription immediately and
create a new one with the same subscription number, new quantity, and
the previous end date for the initial subscription.  Prior to this fix,
the code would perform this segmentation every time a sync ran without
checking whether the subscription had already been segmented.  The
result was an ever increasing number of entries for the same
subscription number.

This patch changes the sort order to DESC when collecting records for a
subscription so that we always operate on the active subscription rather
than a segment that has been terminated.  As well, this patch no longer
considers start date when matching against results from the subscription
service.  Instead, we match only against the subscription number.  The
subscription start date should not change in the service and should be
unique (i.e. we shouldn't see two subscriptions from the service with
the same subscription number but different start dates).