SWATCH-2583: Do not remove unseen finest granularity measurements
The CombiningRollupStrategy relies on a DateRange to lookup and catalog
the existing affected snapshots when creating/updating the finest granularity
snapshots. BEFORE the new tally changes, the strategy could always make the
assumption that all Events in that range were (re)pulled and calculated, and
if an existing snapshot does not have a calculation based on an Event,
the tally measurements could be cleared. The new tally changes, however,
does not make that guarantee. Before producing snapshots, it determines the
affected synthetic range based on the produced account calcs (start from the oldest,
end on the newest). Because there may not be a calc for a date in the middle
of this range, it assumes that the affected snapshots will be everything in
between, and removes the tally measurements for any snaps not represented by
a calculation.

This patch removes the logic that removes the measurements of the finest granularity
snapshots when a calculation did not occur on a specific hour. This is
justified since there was no change to that specific hour.