Support tallying of a hosts that are both PAYG and non-PAYG
Adds support for optional field, addtionalTags to variant section of a swatch product confiurations yaml. Allows us to force additional tags to be applied when a variant is detected. This is what is allowing us to support the business use case where HBI is reporting a traditional base RHEL system, but we want to meter ELS addon applied to the same host. This is ELS specific now, but theoretically could be used to any kind of "addon" scenario where we want to meter the addon, but traditionally count the rhel system.

This approach is flexible and working, but really should be iterated over in the future, perhaps actually implementing logic around the parentSubscription field that's been unused, but the java doc indicated it would be used as part of these ELS epics.