SWATCH-2647: swatch-metrics creates events for els/payg/unconverted
During metric collection, SubscriptionDefinition.getAllProductTagsByRoleOrEngIds returns `RHEL for x86` and `rhel-for-x86-els-payg-addon`.

There was a condition in the code that said if the productTags != 1, skip creating the event.

Updated the logic to just make sure the product tag we're collecting metrics for appears in the tags returned by getAllProductTagsByRoleOrEngIds.

A more elegant solution will be part of SWATCH-2634.