SWATCH-2521: Remove use of subscription_product_ids table
At this time, subscription contains the required fields from subscription_product_ids. Data can be obtained from the subscriptions table by:

- sku->sku_product_tag equivalent to subscription_product_ids>product_id

Note that I haven't dropped the table because it would be too dangerous (we couldn't revert this change if we spot something). So, I will create a follow-up JIRA task for this if the changes look good.