Group RemittanceSummaryProjections by hardware measurement type
When hardware measurement type was added to the remittance table,
we did not update the get remittance totals projection to include
it. Because of this, the remittance totals query was returning the
remittance total matching remittance records with similar criteria
resulting in 0 required remittance (no record created).

This patch adds hardware measurement type to the projection and the
supporting query filter so that the correct totals are returned.

Co-authored-by: Michael Stead <mstead@redhat.com>