SWATCH-2424: Limit the snapshots to purge in batches
I checked the production database and it seems that there are an huge number (we're speaking about more of millions for every month up to 2022) of snapshots that is being tried to be deleted in a single delete query. 

So, we can't split up the delete query to be done by org_id, because we would have the same problem. The only trustable approach is to limit the query by the number of snapshots. This can be defined using a property and by default is 1000000. 

I could test these changes with such amount of data locally, but yes using the stage data. 

Note that to use this limit, I had to turn the delete query into a native query.