Use a service name that won't be shadowed by Clowder
Clowder creates a service with the same name as the pod label.  This
service is inflexible in that Clowder doesn't allow you to expose any
ports that aren't the public, private, and metrics ports defined in the
ClowdEnv.  We need to expose port 9090, but if we name our service
swatch-mock-prometheus-service, Clowder will clobber it.  So we have to
pick a different name.