Correct errors with keystore references in swatch-contracts
The Quarkus Rest Client code requires a keystore or truststore reference
to begin with either "file:" or "classpath:".  (See
io.quarkus.rest.client.reactive.runtime.RestClientCDIDelegateBuilder#locateStream).
This commit changes the references to the keystores to have the correct
prefix.

Additionally, it explicity sets the truststore to empty.  In production
and stage, all services except UMB should use the system truststore.

This patch also endeavors to create a standard nomenclature for the
references: *_PATH for unprefixed file paths and *_RESOURCE for prefixed
paths.