Fix flaky test involving timestamp format
By applying the OffsetDateParamConverterProvider, it forces the query
param for timestamps to always consistently drop 0, For example,

> 2024-08-16T16:32:47.61154261Z

vs.

> 2024-08-16T16:32:47.611542610Z

I also updated the test to use a fixed timestamp that is formatted
differently depending on whether `OffsetDateTime.toString` or the param
converter provider is used, so that the test will consistently fail if
there is a regression.