Truncate start/end dates when syncing contracts from IT Partner gateway
The partner gateway uses nanosecond precision timestamps. Postgres/HSQL only support microsecond precision, so there is a loss of precision when the timestamps are persisted. To see a couple of examples of postgres having less precision, you can try in the database:
```
select '2024-08-24T12:00:00.000000501Z'::timestamptz
select '2024-08-24T12:00:00.000000500Z'::timestamptz
```
Notice that they produce different values.
This is causing a number of side-effects, namely, the equals methods on contracts and subscriptions are treating records as unequal when they shouldn't and, when we use the startDate as a key as seen [here](https://github.com/RedHatInsights/rhsm-subscriptions/blob/0da69bb0a921fedb8e7aadea2e0b5cd0b1027729/swatch-contracts/src/main/java/com/redhat/swatch/contract/service/ContractService.java#L290), it causes a miss in the mapping to existing records.

With these changes, we are truncating the start and end dates to use the same precision than stored in database.