SWATCH-2870: Configure rest-client timeouts in swatch-contracts
When consuming the Subscription API in the monolith, there were no timeout. Now, in Quarkus, the default timeout is 30 seconds. 
I think we should keep using a timeout, but 30 seconds is too low. 
As part of these changes, I configured it with a timeout of 120 seconds (2 minutes). 
Also, we need to configure the kafka consumer commit strategy to not reject messages if we're still within the acceptable window time.