SWATCH-2813: Use specific clientId's credential in swatch-producer-azure
As of ITPART-1180, the clientId associated to a given Azure subscription is available in the /partnerEntitlements response.
When swatch-producer-azure looks up azure usage context and the context includes a clientId value, it should immediately use the azure client configured with the value, and not loop through any other clientId's client instances.