Improve speed of SubscriptionDefinition lookups
This commit adds a caching component to the SubscriptionDefinition class
to improve the efficiency of looking up SubscriptionDefinition's by
role, tag, and engineering ID.  I've made several design choices in this
commit that I will explain below.

Why Caffeine?  Guava does offer a Cache implementation and we are
already using Guava in this project.  However, even the Guava JavaDocs
explicitly state that Caffeine is the superior choice.  The Caffeine API
is almost identical to Guava's so switching between the two is mostly
just a matter of changing import statements.

Why do we need a scheduler?  Caffeine ordinarily performs its
maintenance and evictions during writes and occasionally after read
operations if writes are rare.  Our cache has an extremely high ratio of
reads to writes.  In that case, the documentation suggests providing the
cache with the Java 9 system-wide scheduling thread so cleanUp
operations can be scheduled.

Why are we using a time based eviction strategy instead of one based on
reference count?  I did a lot of experimenting on this.  The issue is
that while SubscriptionDefinition objects are long lived (and thus using
a weak reference as the cache value and allowing garbage collection to
be the de facto eviction method) in the SubscriptionDefinitionRegistry
class, Optional<SubscriptionDefinition> objects are created ad hoc in
our search methods.  I added an eviction listener, wrote multiple test
cases with System.gc() calls, and could see our entries getting evicted
once the test finished.

Why don't we use SubscriptionDefinition as our cache values?  Our cache
loading method returns an Optional<SubscriptionDefinition>.  While we
can do a .orElse(null) on the optional to resolve the Optional, doing so
decreases the performance of the method.  In my benchmark, the time
required when returning a SubscriptionDefinition by resolving the
Optional increased the run time from about 1.5 seconds to 5.5 seconds.

Why don't we have the cache loading method return a
SubscriptionDefinition or null?  That approach just moves the problem to
a different location.  We're using the Stream API to move through the
SubscriptionDefinition's Variants.  Any Collector is going to either
return an Optional or a collection that we would then have to crack open
and convert an empty Optional/collection to null.