Resize resource requests on curl containers.
The curl containers do not need nearly as much memory and CPU as we were
allocating them.

We can get a better idea of the memory requirements by using the
following PromQL:

```
1.5 * (
  avg(
    quantile_over_time(.99, container_memory_working_set_bytes{namespace='rhsm-prod', container!="POD",container!="", container=~'swatch-billable-usage-purge-remittances|swatch-billable-usage-retry-remittances|swatch-billable-usage-sync|swatch-contracts-offering-sync|swatch-contracts-subscription-sync|swatch-metrics-rhel-sync|swatch-metrics-sync|swatch-system-conduit-sync|swatch-tally-hourly|swatch-tally-purge|swatch-tally-purge-events|swatch-tally-tally'}[60d])
  ) by (container) > 0) / 1000000
```

This query calculates the memory use of the 99th percentile of a job run
and then adds 50% for breathing room.  For most of our jobs, 2Mb was
ample but I've made it an even 10Mb for the request just to be extra
cautious.

The CPU determination is more of an educated guess.  As outlined in
https://blog.kubecost.com/blog/requests-and-limits/, the best way to get
empirical data is to create a recording rule:

```
avg(irate(container_cpu_usage_seconds_total{namespace="rhsm-prod", container!="POD", container!=""}[5m])) by (container)
```

And then use the series gather by that rule like so

```
1.5 * quantile_over_time(.99,container_cpu_usage_irate[7d])
```

We don't have that rule in place, however, so I'm making a conservative
guess of 100 millicores.  The speed of the job container's completion is
not highly critical.  The vast majority of the run time is going to be
curl waiting on the HTTP response.  We mostly just need the curl call to
pass or fail in anything vaguely resembling a reasonable time.

See SWATCH-2736