SWATCH-2910: Filter snapshots by supported metric in the product
When existing snapshots have metrics that are not supported by the project (for example: the product "rhel-for-x86-els-payg-addon" does not support the metrics "Sockets" or "Cores"), 
Then we were creating invalid billable usage remittance entries that were kept indefinetely on pending state.

Also, we are updating the aws and azure producers to properly update these records with status failed.