SWATCH-2887: Delete host tally buckets after processing event batches
After the refactor of the event ingestion, we are handling events in batches ordered only by record date. 

Then, when we handle the events for the same host and the buckets need to be either updated/added/deleted in different transactions, hibernate throws an exception because the bucket version to update or delete does not match with the expected (the bucket version has been updated in other transaction).

Therefore, we are adding a transient "stale" field in the host tally buckets, so we can delete them after processing the batch of events. This way when an event is handled, a previously deleted tally bucket can be "restored" preserving the version.