SWATCH-2915 Specify BYOS aws product code
Our previous implementation was based off the incorrect assumption that the presence of an aws billing code on a system fact indicated that system was tied to a marketplace subscription.  We've since learned that any system provisioned in aws, regardless of if they're paying for the subscription through aws, gets an aws billing code associated with it.

Systems that were provisioned with a red hat gold image get specifically *bp-63a5400a* as a billing product code.  PM identified this product code as indicating BYOS.  If a billing product code is absent, that also indicates BYOS.  A meeting with PM was had, and it was decided that we would maintain a list of product codes to consider BYOS instead of the opposite.  If a billing product code is present and it doesn't appear in the list of BYOS product codes, then consider it a Marketplace system.  Similar updates will need to happen with other system reporters since HBI doesn't centralize any normalization.