SWATCH-2960: Increase usage hourly consumer timeout and context retries
We increased the contracts API timeout to 2 minutes, but we missed to also increase the consumer topic timeout, so we're seeing errors like the following in stage:

```
SRMSG18231: The record 20723 from topic-partition 'platform.rhsm-subscriptions.billable-usage-hourly-aggregate-10' has waited for 62 seconds to be acknowledged. This waiting time is greater than the configured threshold (60000 ms). At the moment 1 messages from this partition are awaiting acknowledgement. The last committed offset for this partition was 20722. This error is due to a potential issue in the application which does not acknowledged the records in a timely fashion. The connector cannot commit as a record processing has not completed.
```

Moreover, I still see some errors getting the context (for aws and azure), I will keep investigate these failures, though I will also increase the retry to 3 (before, we were not retrying).