Create a swatch-unleash-importer clowdapp
The swatch-unleash-importer clowdapp defines a job
that will load an unleash instance with the defined
feature flags used by the subscription watch components.

This work had been previously done by awood, however, this patch
makes some tweaks to allow the image to be built in the
standard quay repo and gets it deployed automatically into the
ephemeral environment.