SWATCH-3006: Support back uom in the events ingestion
The uom property was fully removed as part of SWATCH-1233, but this was causing issues in production because the cost-management service was still emitting events using the removed uom property (reported by SWATCH-3006).

Therefore, the uom support was re-added only for ingesting the events and COST-5575 was reported for the cost-management to use the metric_id property.