SWATCH-2951: Allow to filter by metrics in the Instances Table API
The payg and non payg views are aggregating all the metrics into a single row in jsonb format. And the problem is that we can't easily filter and order using these metrics because it is in jsonb.

So, instead of dealing with the metrics jsonb column, we're now adding a left join to the measurements table (for payg is "instance_monthly_totals" and for non payg is "instance_measurements") - note about performance: that these tables are indexed.

Therefore, we're now using this relationship to filter and also to order the result set. These relationships will not be used when mapping the query result set to the DTO.

Changes
- Removed the distinct in the query since it wasn't necessary (the distinct is done in the view).
- Implement the order using the Hibernate Criteria API because the Spring Specification API can only order by existing fields in the entity (and we now want to also order by programmatically added columns)
- Don't use the Spring Pageable API since the sort is not supported anylonger (if provided, then the programmatically order by is ignored)