SWATCH-3036: Fix pagination to process remittances retries
When asking for next pages, since at the same time, we're also updating the remittances to clear the retry_after, we won't be consuming all the remittances. 

To address this issue, we'll always asking for the first page which eventually reach zero. Moreover, we will first count the number of remittances and won't process more than this count (this is to avoid edge conditions and process new usages that might lead in endless