SWATCH-3041: Populate traceresponse header in spring boot services
Added the "traceresponse" header as similarly as done in Quarkus services: https://github.com/RedHatInsights/rhsm-subscriptions/pull/3660

I also spotted that we were not populating the tracing context in spring boot, to fix this, I had to add the `micrometer-tracing-bridge-otel` and `opentelemetry-api`. 

With these two dependencies, we can now get access to the span context, and the trace ID/span ID will be populated when exporting the logs into splunk.