SWATCH-3049: Enable observation in Spring Boot services for Kafka
- Started a playbook to deploy splunk and all the dependant services locally in order to see how the trace is propagated

For now, I've included two integrations:
(1) swatch-metrics <-> swatch-tally: this integration won't propagate the tracing context because we're consuming the events in batches.
(2) swatch-tally <-> swatch-billable-usage: with these changes, the same trace ID will be used along all the tally hourly operation.

- Added new SPLUNK_DISABLE_CERTIFICATE_VALIDATION property to disable the certificate validation when posting logs into splunk. This is the same property as used in Quarkus services.
- Configure Spring Boot services to enable observation in Kafka.

Note that this must be configured individually for each KafkaTemplate and KafkaListener. For now, I've enabled only the KafkaTemplate for tasks and tally-summary.