Monitor tally values via micrometer
The tally counters are tagged by product, metric ID, and billing
provider.  The counter should only be incremented at the finest
granularity level to prevent over-counting.  Additionally the counter
should not be incremented on "superset" type tally snapshots that have
the _ANY value for usage or service level or for TOTAL type tally
measurements.

See SWATCH-2299