Handle unresolved Clowder truststore references better
Truststore handling for the Clowder configuration is complex.  The
Clowder json looks like:

```
"endpoints": [
  {
    "app": "foo",
    "name": "service",
    "hostname": "localhost",
    "tlsPort": 9443
  }
],
"tlsCAPath": "/tmp/ca.pem"
}
```

However, our code generally doesn't want just a bare PEM file.
Instead, we take that path and construct a temporary Java truststore (a
PKCS12 file) from it.  We pass the reference to that keystore along.

Following the pattern of clowder-quarkus-config-source, we also
configure the truststore information on a per-endpoint basis.  To do
that, *if and only if* the endpoint has a `tlsPort` key, then we
generate three synthetic properties: trust-store-path,
trust-store-password, and trust-store-type.  We populate those synthetic
properties with the information from the generated PKS12.

However, if an endpoint doesn't have a `tlsPort` defined on it or if the
`tlsCAPath` isn't specified, the ClowderJsonPropertyResolver can't
resolve the synthetic `clowder.endpoints.[...].trust-store-path`
property. The Spring property resolver behavior is to just return the
literal string instead.

This commit documents this behavior and modifies the CertInfoContributor
to better reflect what is actually going on with the configuration
handling.