Do not index the files in the bin directories
Intellij creates bin directories with compile class files and copied
resource files.  The issue is that when you run a find in files command,
the copied resource files in the bin directories show up in the results.

Also added a directive to tell Intellij to download the source for
libraries.