SWATCH-3066: Print traces for "com.redhat.swatch" in spring services
When investaging the random failure of the test "test_export_instances_payg_product_filters" in SWATCH-3066, I found that we're only seeing the WARN traces when handling the export requests. 

This is because we moved the export handling into a separated module that uses the package "com.redhat.swatch" and the spring boot services are only configured to print the info traces for the package "org.candlepin".

For the Quarkus services, this is already working fine because it's configured as:

```
quarkus.log.level=${LOGGING_LEVEL_ROOT}
quarkus.log.category."com.redhat.swatch".level=${LOGGING_LEVEL_COM_REDHAT_SWATCH}
```

After these changes, the package "com.redhat.swatch" should also be configured for the spring services to print the info traces and see them in splunk.