Add event filtering for incoming HBI events
When HBI event messages are received, filter them to ensure
that we do not process any that define a billing_model of
'marketplace' or a host_type of 'edge'.

This filter mimics the query currently being made to fetch
applicable host records from HBI, with the exception of
orgId.

Since the query (used by the nightly) is made based on the
opted-in org list, we need to also filter any incoming
Swatch Event messages so that Events are not persisted for
orgs that will not be tallied. We do not currently want to
open upnightly tally for every org that sends data to HBI.

NOTE:
The opt-in check added to the EventController does not
add any overhead since we continue to do at most one check
per event.