SWATCH-3050: Update usage context subscriptions counters by product
These changes are about to update the existing counters when getting the subscription usage for aws and azure:
- swatch_missing_subscriptions
- swatch_ambiguous_subscriptions

By adding the "product" dimension.
After merging these changes, we could add a plot to see the numbers by product to early detect misconfigurations.

Apart from these changes, I converted the existing UsageContextSubscriptionProviderTest to a QuarkusTest, to better cover the SQL query.