Add a primary key to all tables that lack one
For logical replication, Postgres requires that all tables have a
primary key.  This commit adds primary keys to tables like the
databasechangelog and join tables.  It also adds the
"pg_stat_statements" library to our database container so that we can
monitor queries in development.

"A published table must have a 'replica identity' configured in order to
be able to replicate UPDATE and DELETE operations, so that appropriate
rows to update or delete can be identified on the subscriber side. By
default, this is the primary key, if there is one."

From https://www.postgresql.org/docs/12/logical-replication-publication.html