Determining remittance total no longer considers hardware_type
When determining the current remittance total, the billable usage
component no longer considers the hardware_measurement_type. As
such, when TallySummary messages are created during the hourly tally,
we must also exclude the measurement type when calculating the
current total for a metric, as this value is used when determining
remittance.