SWATCH-3134: Adding JVM config to disable the trimming of stack traces
The “OmitStackTraceInFastThrow” was first introduced in JDK5. According to the release note of JDK5, the compiler in the server VM stops outputting the stack trace of the exceptions which have been thrown a few times. This is a one of optimizations which happens in C2 compile. And by adding “-XX:-OmitStackTraceInFastThrow”, we can disable this optimization.

This seems to be the cause why we're not getting the full stacktraces in production.