SWATCH-3211: Billable Usage service validate supported metric by product
After SWATCH-2910, we are not creating tally snapshots for unsupported metrics (not existing metrics or metrics not supported by product).

However, when programmatically sending tally snapshots in manual testing or in component tests, this scenario makes the billable usage service to fail and enter in a crash loop exceptions state.

With these changes, we prevent remittances to be created for invalid snapshots and configure the tally snapshot consumer to ignore failures.