SWATCH-3211: Ignore failures when consuming tally summary messages
Using `mp.messaging.incoming.tally-summary.failure-strategy=ignore` will prevent the service to enter in a crash loop exceptions state when consuming invalid messages.

When programmatically sending tally snapshots in manual testing or in component tests, this scenario makes the billable usage service to fail and enter in a crash loop exceptions state.

With these changes, we prevent remittances to be created for invalid snapshots and configure the tally snapshot consumer to ignore failures.