SWATCH-3294: Rename "swatch_billable_usage_total" to aggregated
This is caused because the tickets SWATCH-2301 and SWATCH-2304 were about to create the same metric in two different places, but the changes in SWATCH-2304 didn't include an special status "aggregated", so we're couting the same metric twice with the same values.

After speaking with Kevin Howell, we aggreed to replace the one from SWATCH-2304 to use a new metric counter named "swatch_billable_usage_total_aggregated" to avoid confusions.